refactor(cardano-services): consolidate `ProgramsCommon` with `Program`
