+++ packages/cardano-services/src/Program/errors/index.ts
@@ -1,3 +1,4 @@
 export * from './MissingProgramOption';
 export * from './MissingCardanoNodeOption';
 export * from './UnknownServiceName';
+export * from './WrongOption';
