+++ packages/cardano-services/src/Program/index.ts
@@ -2,7 +2,8 @@ export * from './defaults';
 export * from './errors';
 export * from './loadHttpServer';
 export * from './loadTxWorker';
-export * from './ProgramOptionDescriptions';
+export * from './Options';
+export * from './programs';
 export * from './ServiceNames';
 export * from './utils';
 export * from './services';
