+++ packages/cardano-services/src/Program/services/ogmios.ts
@@ -1,7 +1,7 @@
 /* eslint-disable max-len */
 /* eslint-disable promise/no-nesting */
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { CommonOptionDescriptions, CommonProgramOptions } from '../../ProgramsCommon';
+import { CommonOptionDescriptions, CommonProgramOptions } from '../Options';
 import { DnsResolver } from '../utils';
 import { Logger } from 'ts-log';
 import { MissingCardanoNodeOption } from '../errors';
