--- packages/cardano-services/src/ProgramsCommon/CommonOptionDescriptions.ts
@@ -1,10 +1,0 @@
-export enum CommonOptionDescriptions {
-  DbCacheTtl = 'Cache TTL in minutes between 1 and 2880, an option for database related operations',
-  LoggerMinSeverity = 'Log level',
-  OgmiosSrvServiceName = 'Ogmios SRV service name',
-  OgmiosUrl = 'Ogmios URL',
-  RabbitMQSrvServiceName = 'RabbitMQ SRV service name',
-  RabbitMQUrl = 'RabbitMQ URL',
-  ServiceDiscoveryBackoffFactor = 'Exponential backoff factor for service discovery',
-  ServiceDiscoveryTimeout = 'Timeout for service discovery attempts'
-}
