--- packages/cardano-services/src/ProgramsCommon/defaults.ts
@@ -1,12 +1,0 @@
-import { Ogmios } from '@cardano-sdk/ogmios';
-
-export const OGMIOS_URL_DEFAULT = (() => {
-  const connection = Ogmios.createConnectionObject();
-  return connection.address.webSocket;
-})();
-
-export const RABBITMQ_URL_DEFAULT = 'amqp://localhost:5672';
-
-export const USE_QUEUE_DEFAULT = false;
-
-export const ENABLE_METRICS_DEFAULT = false;
