--- packages/cardano-services/src/ProgramsCommon/errors/index.ts
@@ -1,1 +1,0 @@
-export * from './WrongOption';
