--- packages/cardano-services/src/ProgramsCommon/index.ts
@@ -1,5 +1,0 @@
-export * from './CommonOptionDescriptions';
-export * from './defaults';
-export * from './errors';
-export * from './options';
-export * from './programs';
