--- packages/cardano-services/src/ProgramsCommon/options.ts
@@ -1,16 +1,0 @@
-import { LogLevel } from 'bunyan';
-
-/**
- * Common options for programs:
- * - HTTP server
- * - RabbitMQ worker
- */
-export interface CommonProgramOptions {
-  loggerMinSeverity?: LogLevel;
-  ogmiosUrl?: URL;
-  rabbitmqUrl?: URL;
-  ogmiosSrvServiceName?: string;
-  rabbitmqSrvServiceName?: string;
-  serviceDiscoveryBackoffFactor?: number;
-  serviceDiscoveryTimeout?: number;
-}
