+++ packages/cardano-services/src/cli.ts
@@ -6,30 +6,29 @@
 /* eslint-disable unicorn/no-nested-ternary */
 import {
   API_URL_DEFAULT,
+  CommonOptionDescriptions,
+  ENABLE_METRICS_DEFAULT,
   HttpServerOptions,
   OGMIOS_URL_DEFAULT,
   PARALLEL_MODE_DEFAULT,
   PARALLEL_TXS_DEFAULT,
   POLLING_CYCLE_DEFAULT,
   ProgramOptionDescriptions,
+  Programs,
   RABBITMQ_URL_DEFAULT,
   SERVICE_DISCOVERY_BACKOFF_FACTOR_DEFAULT,
   SERVICE_DISCOVERY_TIMEOUT_DEFAULT,
   ServiceNames,
   TxWorkerOptionDescriptions,
   TxWorkerOptions,
+  USE_QUEUE_DEFAULT,
+  WrongOption,
   loadAndStartTxWorker,
   loadHttpServer,
   loadSecret
 } from './Program';
 import { Command, Option } from 'commander';
-import {
-  CommonOptionDescriptions,
-  ENABLE_METRICS_DEFAULT,
-  Programs,
-  USE_QUEUE_DEFAULT,
-  WrongOption
-} from './ProgramsCommon';
+
 import { DB_CACHE_TTL_DEFAULT } from './InMemoryCache';
 import { DEFAULT_TOKEN_METADATA_CACHE_TTL, DEFAULT_TOKEN_METADATA_SERVER_URL } from './Asset';
 import { EPOCH_POLL_INTERVAL_DEFAULT } from './util';
